Moved unit tests into own subdir
