+++ __tests__/unit/InputRulesBuilder.test.ts
@@ -1,11 +1,11 @@
 import { mocked } from "jest-mock";
 import { Plugin } from "prosemirror-state";
 
-import { ExtensionManager } from "../src/ExtensionManager";
-import { KeymapBuilder } from "../src/KeymapBuilder";
-import { MockNodeExtension } from "./mocks/MockNodeExtension";
+import { ExtensionManager } from "../../src/ExtensionManager";
+import { InputRulesBuilder } from "../../src/InputRulesBuilder";
+import { MockNodeExtension } from "../mocks/MockNodeExtension";
 
-test("KeymapBuilder creates a plugin", () => {
+test("InputRulesBuilder creates a plugin", () => {
   const docExtension = mocked(new MockNodeExtension());
   docExtension.proseMirrorNodeName.mockReturnValueOnce("doc");
   docExtension.proseMirrorNodeSpec.mockReturnValueOnce({});
@@ -15,7 +15,7 @@ test("KeymapBuilder creates a plugin", () => {
 
   const manager = new ExtensionManager([docExtension, textExtension]);
 
-  const builder = new KeymapBuilder(manager);
+  const builder = new InputRulesBuilder(manager);
   const rules = builder.build();
 
   expect(rules).toBeInstanceOf(Plugin);
