+++ __tests__/unit/MarkExtension.test.ts
@@ -1,9 +1,9 @@
 import { Schema } from "prosemirror-model";
 
-import { Extension } from "../src/Extension";
-import { MarkExtension } from "../src/MarkExtension";
-import { SyntaxExtension } from "../src/SyntaxExtension";
-import { MockMarkExtension } from "./mocks/MockMarkExtension";
+import { Extension } from "../../src/Extension";
+import { MarkExtension } from "../../src/MarkExtension";
+import { SyntaxExtension } from "../../src/SyntaxExtension";
+import { MockMarkExtension } from "../mocks/MockMarkExtension";
 
 test("MarkExtension.proseMirrorToUnistTest has a default implementation", () => {
   const extension = new MockMarkExtension();
